Do not allow reading/flashing when radio is detected to be turned on/connected.
Use MSP_STATUS to detect failsafe state. If RX_FAILSAFE is NOT detected, it means that the radio is on and connected.
Due to changes in BF 4.3.x and moving forward issues might arise - especially with SPI receivers - where the communication might get flaky. This seems to be due to changes in the scheduler and interrupt handling. This may lead to corrupted flashes resulting in soft or full brick of the ESCs (recovery only possible via C2 interface).
This is considered to be a workaround. Other functionality where BF communicates with the ESC such as for example direction changes are also affected, but obviously there is nothing that can be done about that in esc-configurator.